Add files via upload
Sorry for any hard edges or anything; this is my first go, and I hope to expand upon things; but this will be just the start and I shall update as I code new things in; I figured I would post everything here in case anyone feels the need to point anything out or fork it. Thank you all for looking!